Convert price from long to double KAN-96

+++ backend/src/main/kotlin/ch/zhaw/pm4/simonsays/api/types/MenuItemDTO.kt
@@ -10,5 +10,5 @@ data class MenuItemDTO (
         @field:Schema(description = "List of ingredients")
         var ingredients: List<IngredientDTO>,
         @field:Schema(description = "Price of the menu item")
-        var price: Long
+        var price: Double
 )
